Renamed module to autoria.py
