+++ average_calculation.py
@@ -5,16 +5,16 @@ average used car prices that are sold on http://auto.ria.com
 """
 
 from pprint import pprint
-from api_methods import get_average
+from autoria import get_average
 
 PARAMETERS = {
     'VEHICLE_TYPE': 'Легковые',
-    'VEHICLE_MARK': 'Ford',
-    'VEHICLE_MODEL': 'Focus',
+    'VEHICLE_MARK': 'Renaul',
+    'VEHICLE_MODEL': 'Scenic',
     'VEHICLE_BODY': None,
-    'VEHICLE_YEAR_START': 2000,
+    'VEHICLE_YEAR_START': 2005,
     'VEHICLE_YEAR_END': 2006,
-    'VEHICLE_REGION': 'Харьков',
+    'VEHICLE_REGION': None,
     # not implemented yet
     # VEHICLE_MIN_PRICE = '4000'
     # VEHICLE_MAX_PRICE = '5000'
